Refactor BDAddr into submodule
